Got the main file ported over, gotta make it work for generator-dispatcher
Move the ConsumerConfig to kafka lib

Move context to its own file

formatting
