+++ src/rust/generator-dispatcher/Cargo.toml
@@ -23,5 +23,18 @@ tokio = { version = "1.17", features = ["macros", "rt", "rt-multi-thread"] }
 tracing = "0.1.29"
 uuid = { version = "0.8", features = ["v4"] }
 
+[dev-dependencies]
+async-trait = "0.1"
+bytes = "1.0"
+opentelemetry = { version = "0.16", features = ["rt-tokio"] }
+opentelemetry-jaeger = { version = "0.15", features = [
+  "collector_client",
+  "rt-tokio"
+] }
+test-context = "0.1"
+tracing-appender = "0.2"
+tracing-opentelemetry = "0.16"
+tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
+
 [features]
 new_integration_tests = []
