+++ src/rust/generator-dispatcher/src/main.rs
@@ -1,9 +1,6 @@
 use clap::StructOpt;
 use futures::StreamExt;
-use generator_dispatcher::config::{
-    build_consumer,
-    GeneratorDispatcherConfig,
-};
+use generator_dispatcher::config::GeneratorDispatcherConfig;
 use kafka::Consumer;
 use plugin_work_queue::client::{
     FromEnv,
@@ -39,7 +36,7 @@ impl GeneratorDispatcher {
         let plugin_work_queue_client = PluginWorkQueueServiceClient::from_env().await?;
 
         let raw_logs_consumer: Consumer<Envelope<RawLog>> =
-            build_consumer(config.kafka_config, "raw-logs".to_string())?;
+            Consumer::new_from_config(config.kafka_config, "raw-logs".to_string())?;
 
         Ok(Self {
             plugin_work_queue_client,
