+++ src/rust/generators/sysmon-generator/Cargo.toml
@@ -6,6 +6,11 @@ edition = "2021"
 
 [dependencies]
 chrono = { version = "0.4" }
+clap = { version = "3.0", default_features = false, features = [
+  "std",
+  "env",
+  "derive"
+] }
 endpoint-plugin = { path = "../../endpoint-plugin", version = "*" }
 futures = "0.3"
 kafka = { path = "../../kafka", version = "*" }
