+++ src/rust/kafka/Cargo.toml
@@ -9,6 +9,11 @@ license = "MIT"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
 bytes = "1.1"
+clap = { version = "3.0", default_features = false, features = [
+  "std",
+  "env",
+  "derive"
+] }
 futures = "0.3"
 rdkafka = { version = "0.28", features = [
   "cmake-build",
